Make post SDK startup init phases independently failable (#3934)
* Move SDK init methods

* Made SDK post startup init phase failures independent
