+++ embrace-android-infra/src/main/kotlin/io/embrace/android/embracesdk/internal/logging/InternalErrorType.kt
@@ -47,4 +47,5 @@ sealed class InternalErrorType(private val severity: Severity) {
     object NavControllerTrackingFail : InternalErrorType(ERROR)
     object UserSessionCallbackFail : InternalErrorType(ERROR)
     object HttpRequestInfoModifierFail : InternalErrorType(Severity.WARNING)
+    object SdkInitPhaseFail : InternalErrorType(ERROR)
 }
